# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TogetherPlan API",
        "description": "API documentation for the TogetherPlan project. This API allows users to manage events, participants, voting on dates, groups, and notifications.",
        "contact": {
            "name": "TogetherPlan Support",
            "email": "support@togetherplan.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.togetherplan.com",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of events",
                "description": "List all public events or private events if the user is invited",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "parameters": [
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter events by location",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Casablanca"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter events by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Meeting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Event"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Store a new event",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "category",
                                    "privacy"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Team Meeting"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Discuss project progress"
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Casablanca"
                                    },
                                    "category": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Meeting"
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ],
                                        "example": "private"
                                    },
                                    "date_options": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "proposed_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-11-15"
                                                },
                                                "proposed_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "example": "14:30:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "invitees": {
                                        "description": "Array of user IDs to invite",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get details of a single event",
                "description": "Show event details",
                "operationId": "6f743f71f7a88dbe57be1d9c9ebf47cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to view this event"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an existing event",
                "description": "Update an event",
                "operationId": "fb130dcf1ab4d7abbd3a173741df1f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "category": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ]
                                    },
                                    "group_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to update this event"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "description": "Delete an event",
                "operationId": "b8c8f4b41a32532f5e9f7146e511bc61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized to delete this event"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "title": "Event",
                "description": "Schema representing an event",
                "required": [
                    "id",
                    "title",
                    "location",
                    "category",
                    "privacy",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Team Meeting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Discuss project progress"
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "category": {
                        "type": "string",
                        "example": "Meeting"
                    },
                    "privacy": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 3
                    },
                    "group_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "best_date_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "creator": {
                        "$ref": "#/components/schemas/User"
                    },
                    "group": {
                        "$ref": "#/components/schemas/Group"
                    },
                    "participants": {
                        "description": "List of participants for this event",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Participant"
                        }
                    },
                    "dateOptions": {
                        "description": "List of proposed date options for this event",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DateOption"
                        }
                    },
                    "bestDate": {
                        "$ref": "#/components/schemas/DateOption"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    }
                },
                "type": "object"
            },
            "Group": {
                "title": "Group",
                "required": [
                    "id",
                    "name",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Groupe Dev"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du groupe"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Participant": {
                "title": "Participant",
                "description": "Schema representing a participant in an event",
                "properties": {
                    "id": {
                        "description": "ID du participant",
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "description": "ID de l'événement auquel le participant est associé",
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur participant",
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "description": "Statut de l'invitation",
                        "type": "string",
                        "enum": [
                            "invited",
                            "accepted",
                            "declined"
                        ],
                        "example": "invited"
                    },
                    "created_at": {
                        "description": "Date de création du participant",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour du participant",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T16:05:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Schema representing a user",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Zakaria Ghazi"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "zakariaghazi77@gmail.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T15:50:50Z"
                    },
                    "groups": {
                        "description": "Groups the user belongs to",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Group"
                        }
                    },
                    "participants": {
                        "description": "Participant records for this user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Participant"
                        }
                    },
                    "events_created": {
                        "description": "Events created by this user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum API token for authentication",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}